+++ HtmlRun.Runtime/Constants/StringInstructionsSet.cs
@@ -1,0 +1,12 @@
+namespace HtmlRun.Runtime.Constants;
+
+public static class StringInstructionsSet
+{
+  public const string Trim = "Trim";
+  
+  public const string ToLowerCase = "ToLowerCase";
+
+  public const string ToUpperCase = "ToUpperCase";
+
+  public const string ToTitleCase = "ToTitleCase";
+}