+++ HtmlRun.Terminal/Providers/TerminalProvider.cs
@@ -1,0 +1,11 @@
+using HtmlRun.Runtime.Native;
+using HtmlRun.Terminal.Instructions;
+
+namespace HtmlRun.Terminal;
+
+class TerminalProvider : INativeProvider
+{
+  public string Namespace => "Console";
+
+  public INativeInstruction[] Instructions => new INativeInstruction[] { new ReadLineCmd(), };
+}